[ExmoAPI][via CURL][addition]: add rules in xmake.lua

+++ exmo_api/via_curl/src/exmo_api.cpp
@@ -6,7 +6,7 @@ namespace TB_NS::Exmo_NS {
         , m_secret_key(std::move(i_secret_key))
         , m_url("api.exmo.com/v1/")
         , m_connection(HTTP_NS::Connection())
-        , m_nonce(std::time(nullptr)) {}
+        , m_nonce(static_cast<ULONG>(std::time(nullptr))) {}
 
     HTTP_NS::JsonData API::call(std::string_view i_method, std::string_view i_params) {
         std::string params = "nonce=";
